requirements.txt
python-binance
python-telegram-bot==13.15
pandas
gspread
oauth2client
python-dotenv
requests
schedule
